Fix four latent defects found while auditing for simplifications
minkowski_error divided by to_type(power * batch_size). to_type only takes an
Index, so the float product was truncated before being widened back: with the
default power of 1.5, any odd batch size lost the .5. That left the error
inconsistent with its own gradient, which divides by batch_size alone - the
correct derivative of |d|^p/(p*n) only when the divisor is not truncated. A
tree-wide sweep of to_type/to_int with arithmetic arguments confirms this was
the only instance of the pattern.

Both TimeSeriesDataset correlation functions sized their result by one
predicate and filled it by another. role_applies_to puts InputTarget variables
in both the Input and Target lists, so the two rules diverge as soon as types
or ordering vary; calculate_autocorrelations compounded it by using
extra_targets, a count, as a slice length into a list that is not ordered
pure-targets-first. When the sizing came out short the fill loop wrote past
the end. Both now size from the same numeric_variable_indices vector they
iterate, which also removes four locals.

GrowingNeurons printed elapsed_time before computing it, so every epoch
displayed the previous epoch's value; genetic_algorithm already had the right
order. calculate_autocorrelations logged unconditionally while its sibling
honoured the display flag.

Three cuDNN graph sites discarded a [[nodiscard]] error status from
get_workspace_size and now use the value-returning overload. Shared-scratch
growth was recorded under "workspace.cudnn_frontend" even when driven by
cuBLASLt matmul, mis-attributing it in the memory reports the footprint
benchmarks read. ensure_int8_dequant_workspace and
release_matmul_thread_workspaces had no CPU stubs, unlike their seven
siblings, so an unguarded call would fail at link time rather than throw.

The remainder is mechanical: redundant guards and allocations, dead locals and
assignments, a recomputed mean, two shadowing one/zero locals, five no-op
static_casts, a single-use lambda, a duplicated tokenizer loop, duplicate
includes, an unused include, and the SoftmaxFitOptions struct whose five
never-varied fields left two branches unreachable.

Verified on both builds after every change: 864/864 CUDA, 801/803 CPU.

+++ opennn/network_differential.cpp
@@ -64,10 +64,10 @@ void NetworkDifferential::build(const NeuralNetwork& network)
 
             for (Index j = 0; j < features; ++j)
             {
-                snapshot.minimum(j)   = static_cast<float>((*descriptives)[j].minimum);
-                snapshot.maximum(j)   = static_cast<float>((*descriptives)[j].maximum);
-                snapshot.mean(j)      = static_cast<float>((*descriptives)[j].mean);
-                snapshot.deviation(j) = static_cast<float>((*descriptives)[j].standard_deviation);
+                snapshot.minimum(j)   = (*descriptives)[j].minimum;
+                snapshot.maximum(j)   = (*descriptives)[j].maximum;
+                snapshot.mean(j)      = (*descriptives)[j].mean;
+                snapshot.deviation(j) = (*descriptives)[j].standard_deviation;
             }
             break;
         }
